requirements.txt
flask
flask_sqlalchemy
flask_login
openai
python-dotenv
flask-wtf
wtforms
requests
openai>=1.0.0
PyPDF2
Flask
Flask-SQLAlchemy
Flask-Login
openai
python-dotenv
PyPDF2
pandas

